pyproject.toml
[project]
name = "researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "cloudinary>=1.44.1",
    "faiss-cpu>=1.12.0",
    "fastapi[standard]>=0.116.1",
    "grandalf>=0.8",
    "ipykernel>=6.30.1",
    "ipython>=9.5.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.35",
    "langchain-postgres>=0.0.15",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-postgres>=2.0.23",
    "pgvector>=0.3.6",
    "psycopg[binary,pool]>=3.2.9",
    "psycopg2>=2.9.10",
    "pypdf>=6.0.0",
    "pyppeteer>=2.0.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
